feat : removed old zk whir opt
