+++ src/protocols/whir_zk/mod.rs
@@ -1,240 +1,205 @@
+/// zkWHIR 2.0 — Zero-Knowledge WHIR with poly-logarithmic overhead.
+///
+/// Uses the "Alternative Randomness Sampling" approach which samples only
+/// ν = ⌊μ/ℓ⌋ + 1 blinding polynomials (instead of μ + 1), reducing proof
+/// size to (ν + 1) · q(δ) field elements.
+///
+/// Two WHIR instances run as sub-protocols:
+///   1. `blinded_polynomial`: over the μ-variate masked witness f̂ = f + msk(Φ₀)
+///   2. `blinding_polynomial`: over (ℓ+1)-variate committed vectors M and ĝ₁..ĝ_ν
+///
+/// Protocol phases:
+///   Step 1: Commitment — sample msk, ĝ₀..ĝ_ν; commit [[f̂]], [[M]], [[ĝᵢ]]
+///   Step 2: Blinding claims — V samples β; P builds g(x̄) = Σ βⁱ·ĝᵢ(Φᵢ(x̄)), sends G
+///   Step 3: Combination — V samples ρ ≠ 0; P forms f_zk = ρ·f + g
+///   Step 4: Initial sumcheck on f_zk; P sends [[H]] = fold_k(f_zk, r̄)
+///   Step 5: Virtual OOD/STIR queries + remaining WHIR rounds
+///   Step 6: Γ consistency check — verify [[f̂]] openings match [[H]]
+///   Step 7: Batched blinding proof via second WHIR instance
+use ark_ff::FftField;
+use serde::{Deserialize, Serialize};
+
+use crate::algebra::embedding::Embedding;
+
 mod committer;
 mod prover;
 mod utils;
 mod verifier;
 
-use std::fmt::Display;
-
-use ark_ff::FftField;
-use serde::{Deserialize, Serialize};
-
-pub use self::committer::{Commitment, Witness};
+pub use self::{committer::Witness, verifier::Commitments};
 use crate::{
-    algebra::embedding::{Embedding, Identity},
+    algebra::embedding::Identity,
     parameters::ProtocolParameters,
     protocols::{irs_commit, whir},
 };
 
-/// Policy inputs for `ell` computation.
-///
-/// Leakage upper bound used by the current sizing rule:
-/// `q_ub = k1*q_delta_1 + k2*q_delta_2 + (d+1)*mu + t1 + t2`.
-#[derive(Clone, Copy, PartialEq, Eq, Serialize, Deserialize, Debug)]
-pub struct BlindingSizePolicy {
-    /// Query count for witness-side WHIR first-round leakage term.
-    pub q_delta_1: usize,
-    /// Query count for blinding-side WHIR first-round leakage term.
-    pub q_delta_2: usize,
-    /// Last-round clear leakage budget for witness-side WHIR.
-    pub t1: usize,
-    /// Last-round clear leakage budget for blinding-side WHIR.
-    pub t2: usize,
-    /// Degree `d` of each sumcheck round polynomial in the leakage term `(d+1)*mu`.
-    pub sumcheck_round_degree: usize,
-}
-
-impl BlindingSizePolicy {
-    pub fn from_whir_params<F: FftField>(main_whir_params: &ProtocolParameters) -> Self {
-        // TODO: Compute these in a cleaner way?
-
-        let protocol_security_level_main = main_whir_params
-            .security_level
-            .saturating_sub(main_whir_params.pow_bits);
-        #[allow(clippy::cast_possible_wrap)]
-        let q_delta_1 = irs_commit::num_in_domain_queries(
-            main_whir_params.unique_decoding,
-            protocol_security_level_main as f64,
-            0.5_f64.powi(main_whir_params.starting_log_inv_rate as i32),
-        );
-        #[allow(clippy::cast_possible_wrap)]
-        let q_delta_2 = irs_commit::num_in_domain_queries(
-            main_whir_params.unique_decoding,
-            main_whir_params.security_level as f64,
-            0.5_f64.powi(main_whir_params.starting_log_inv_rate as i32),
-        );
-
-        // Default send-in-clear thresholds match query complexities.
-        Self {
-            q_delta_1,
-            q_delta_2,
-            t1: q_delta_1,
-            t2: q_delta_2,
-            // For PCS, round polynomials are cubic (d=3), giving 4*mu.
-            sumcheck_round_degree: 3,
-        }
-    }
-}
-
-/// ZK WHIR configuration: witness-side WHIR + blinding-side WHIR.
+#[allow(clippy::trait_duplication_in_bounds)]
 #[derive(Clone, PartialEq, Eq, Serialize, Deserialize, Debug)]
 #[serde(bound = "")]
 pub struct Config<F: FftField> {
-    pub blinded_commitment: whir::Config<Identity<F>>,
-    pub blinding_commitment: whir::Config<Identity<F>>,
+    /// First WHIR instance: proves claims about f_zk = ρ·f + g over 2^μ evaluations.
+    pub blinded_polynomial: whir::Config<Identity<F>>,
+    /// Second WHIR instance: batched proof of blinding polynomial evaluations
+    /// over 2^(ℓ+1) evaluations with n + ν committed vectors.
+    pub blinding_polynomial: whir::Config<Identity<F>>,
 }
 
 impl<F: FftField> Config<F> {
-    /// Build a zkWHIR config from the given WHIR parameters.
-    pub fn new(size: usize, main_whir_params: &ProtocolParameters, num_polynomials: usize) -> Self {
-        let size_policy = BlindingSizePolicy::from_whir_params::<F>(main_whir_params);
-        Self::new_with_blinding_size_policy(size, main_whir_params, num_polynomials, size_policy)
-    }
+    pub fn new(params: &ProtocolParameters, num_variables_main: usize) -> Self {
+        let blinded_config: whir::Config<Identity<F>> =
+            whir::Config::new(1 << num_variables_main, params);
+        let witness_sec = params.security_level.saturating_sub(params.pow_bits) as f64;
+        let blinding_sec = params.security_level as f64;
 
-    /// Build a zkWHIR config with an explicit blinding size policy.
-    ///
-    /// * `blinding_folding_factor` — folding factor for the blinding-side WHIR.
-    /// * `num_polynomials` — number of polynomials that will be committed.
-    /// * `size_policy` — controls the blinding polynomial size `ell`.
-    pub fn new_with_blinding_size_policy(
-        size: usize,
-        main_whir_params: &ProtocolParameters,
-        num_polynomials: usize,
-        size_policy: BlindingSizePolicy,
-    ) -> Self {
-        let blinded_commitment = whir::Config::new(size, main_whir_params);
-        let num_witness_variables = blinded_commitment.initial_num_variables();
-        let blinding_first_round_interleaving_depth =
-            1usize << main_whir_params.initial_folding_factor;
-        let num_blinding_variables = Self::compute_num_blinding_variables(
-            &blinded_commitment,
-            blinding_first_round_interleaving_depth,
-            size_policy,
+        // T(δ) for the witness instance: the polynomial size sent in the final WHIR round.
+        let witness_t_delta = blinded_config.final_sumcheck.initial_size;
+
+        let mut witness_leak = InstanceLeak::new(
+            params,
+            witness_sec,
+            &blinded_config.initial_committer,
+            num_variables_main,
+        );
+        witness_leak.t_delta = witness_t_delta;
+        // For the blinding instance, T(δ) is approximated as q(δ) since its config
+        // depends on ℓ, which we're computing here (conservative upper bound).
+        let blinding_leak = InstanceLeak::new(
+            params,
+            blinding_sec,
+            &blinded_config.initial_committer,
+            num_variables_main,
         );
 
-        let blinding_size = 1 << num_blinding_variables;
-        let blinding_whir_params = ProtocolParameters {
-            batch_size: num_polynomials * (num_witness_variables + 1),
-            ..*main_whir_params
-        };
-        let blinding_commitment = whir::Config::new(blinding_size, &blinding_whir_params);
+        let q_ub = query_upper_bound(&witness_leak, &blinding_leak);
 
-        Self {
-            blinded_commitment,
-            blinding_commitment,
-        }
-    }
-
-    fn compute_num_blinding_variables(
-        blinded: &whir::Config<Identity<F>>,
-        blinding_first_round_interleaving_depth: usize,
-        size_policy: BlindingSizePolicy,
-    ) -> usize {
-        // Doc formula:
-        // q_ub = k1*q(delta1) + k2*q(delta2) + (d+1)*mu + T1 + T2,
-        // choose smallest ell with 2^ell > q_ub.
-        let num_witness_variables = blinded.initial_num_variables();
+        // ell = smallest integer such that 2^ell > q_ub
+        let ell = (usize::BITS - q_ub.leading_zeros()) as usize;
         assert!(
-            size_policy.t1 >= size_policy.q_delta_1,
-            "invalid blinding size policy: T1 must satisfy T1 >= q(delta1)"
-        );
-        assert!(
-            size_policy.t2 >= size_policy.q_delta_2,
-            "invalid blinding size policy: T2 must satisfy T2 >= q(delta2)"
-        );
-        let k1 = 1usize << blinded.initial_sumcheck.num_rounds;
-        let k2 = blinding_first_round_interleaving_depth;
-        let sumcheck_coeff_leakage = size_policy
-            .sumcheck_round_degree
-            .saturating_add(1)
-            .saturating_mul(num_witness_variables);
-        let query_upper_bound = k1
-            .saturating_mul(size_policy.q_delta_1)
-            .saturating_add(k2.saturating_mul(size_policy.q_delta_2))
-            .saturating_add(size_policy.t1)
-            .saturating_add(size_policy.t2)
-            .saturating_add(sumcheck_coeff_leakage);
-
-        let num_blinding_variables = (usize::BITS - query_upper_bound.leading_zeros()) as usize;
-        assert!(
-            num_blinding_variables < num_witness_variables,
-            "blinding variables ({num_blinding_variables}) must be fewer than \
-             witness variables ({num_witness_variables})"
+            ell + 1 < num_variables_main,
+            "blinding variables ell+1={} must be < mu={num_variables_main}",
+            ell + 1
         );
         debug_assert!(
-            (1usize << num_blinding_variables) > query_upper_bound,
-            "2^ell ({}) must exceed query upper bound ({query_upper_bound})",
-            1usize << num_blinding_variables
+            (1usize << ell) > q_ub,
+            "2^ell ({}) must exceed q_ub ({q_ub})",
+            1usize << ell
         );
-        num_blinding_variables
-    }
 
-    /// Number of variables in the witness polynomial (`μ`).
-    pub fn num_witness_variables(&self) -> usize {
-        self.blinded_commitment.initial_num_variables()
-    }
+        // nu = ⌊mu/ell⌋ — number of blinding polynomials (alternative sampling)
+        let nu = num_variables_main / ell;
+        let blinding_params = ProtocolParameters {
+            batch_size: params.batch_size + nu,
+            ..*params
+        };
 
-    /// Number of blinding variables (`ℓ`).
-    pub fn num_blinding_variables(&self) -> usize {
-        self.blinding_commitment.initial_num_variables() - 1
-    }
-
-    /// Interleaving depth of the initial IRS commitment (= 2^folding_factor).
-    pub(crate) const fn interleaving_depth(&self) -> usize {
-        self.blinded_commitment.initial_committer.interleaving_depth
-    }
-
-    /// Generator ω of the full NTT domain (size = num_rows × interleaving_depth).
-    pub(crate) fn omega_full(&self) -> F {
-        let codeword_length = self.blinded_commitment.initial_committer.codeword_length;
-        let full_domain_size = codeword_length * self.interleaving_depth();
-        crate::algebra::ntt::generator(full_domain_size)
-            .expect("full IRS domain should have primitive root")
-    }
-
-    /// Sub-domain generator (ω_sub = ω^interleaving_depth).
-    fn omega_sub(&self) -> F {
-        self.blinded_commitment.initial_committer.generator()
-    }
-
-    /// ζ = ω^num_rows — the interleaving_depth-th root of unity.
-    pub(crate) fn zeta(&self) -> F {
-        let codeword_length = self.blinded_commitment.initial_committer.codeword_length;
-        self.omega_full().pow([codeword_length as u64])
-    }
-
-    /// Precomputed sub-domain powers [1, ω_sub, ω_sub², ..., ω_sub^(num_rows-1)].
-    pub(crate) fn omega_powers(&self) -> Vec<F> {
-        let codeword_length = self.blinded_commitment.initial_committer.codeword_length;
-        crate::algebra::geometric_sequence(self.omega_sub(), codeword_length)
-    }
-
-    /// Find the index of `alpha_base` in the sub-domain powers.
-    pub(crate) fn query_index(alpha_base: F, omega_powers: &[F]) -> usize {
-        omega_powers
-            .iter()
-            .position(|&p| p == alpha_base)
-            .expect("query point must be in IRS domain")
-    }
-
-    /// Compute all gamma points for a set of query points (flat list).
-    pub(crate) fn all_gammas(&self, query_points: &[F]) -> Vec<F> {
-        let omega_powers = self.omega_powers();
-        let interleaving_depth = self.interleaving_depth();
-        let omega_full = self.omega_full();
-        let zeta_powers = crate::algebra::geometric_sequence(self.zeta(), interleaving_depth);
-        let embedding = self.blinded_commitment.embedding();
-
-        let mut gammas = Vec::with_capacity(query_points.len() * interleaving_depth);
-        for &alpha in query_points {
-            let idx = Self::query_index(alpha, &omega_powers);
-            let coset_offset = omega_full.pow([idx as u64]);
-            for &zp in &zeta_powers {
-                gammas.push(embedding.map(coset_offset * zp));
-            }
+        Self {
+            blinded_polynomial: blinded_config,
+            blinding_polynomial: whir::Config::new(1 << (ell + 1), &blinding_params),
         }
-        gammas
     }
 }
 
-impl<F: FftField> Display for Config<F> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        writeln!(f, "zkWHIR config: witness and blinding commitments")?;
-        writeln!(f, "Witness-side:")?;
-        write!(f, "{}", self.blinded_commitment)?;
-        writeln!(f, "------------------------------------")?;
-        writeln!(f, "Blinding-side:")?;
-        write!(f, "{}", self.blinding_commitment)
+/// Maximum degree of the sumcheck round polynomial.
+///
+/// In each sumcheck round the prover evaluates `eq(r, x) · f(x)` where `eq` is degree 2
+/// and `f` contributes degree 1 from folding, giving a combined degree of 3.
+/// This bounds the (d+1)·μ term in the leakage formula.
+const MAX_SUMCHECK_DEGREE: usize = 3;
+
+/// Per-instance leakage parameters for a single WHIR execution.
+///
+/// `leak(δ, k, μ, d) := k · [q(δ) + stir(δ)] + T(δ) + ood(δ) + (d+1) · μ`
+struct InstanceLeak {
+    /// Folding factor `k = 2^s` for the first round.
+    k: usize,
+    /// Number of sumcheck rounds (= number of witness variables).
+    mu: usize,
+    /// Max degree of sumcheck round polynomial.
+    d: usize,
+    /// `q(δ)`: query complexity.
+    q_delta: usize,
+    /// `stir(δ)`: STIR queries during the first folding round.
+    stir_delta: usize,
+    /// `ood(δ)`: out-of-domain queries.
+    ood_delta: usize,
+    /// `T(δ)`: number of raw coefficients sent during the last sumcheck round.
+    /// Invariant: `T(δ) ≥ q(δ)`.
+    t_delta: usize,
+}
+
+impl InstanceLeak {
+    /// Construct leak parameters for one WHIR instance.
+    ///
+    /// `security_target` is `λ - pow_bits` for the witness side
+    /// or full `λ` for the blinding side.
+    fn new<M>(
+        params: &ProtocolParameters,
+        security_target: f64,
+        irs_config: &irs_commit::Config<M>,
+        num_variables: usize,
+    ) -> Self
+    where
+        M: Embedding,
+        M::Source: FftField,
+        M::Target: FftField,
+    {
+        #[allow(clippy::cast_possible_wrap)]
+        let rate = 0.5_f64.powi(params.starting_log_inv_rate as i32);
+        let (q, stir) = Self::query_counts(
+            params.unique_decoding,
+            security_target,
+            rate,
+            params.initial_folding_factor,
+        );
+
+        Self {
+            k: 1 << params.initial_folding_factor,
+            mu: num_variables,
+            d: MAX_SUMCHECK_DEGREE,
+            q_delta: q,
+            stir_delta: stir,
+            ood_delta: irs_config.out_domain_samples,
+            t_delta: q, // conservative default; overridden for witness instance in Config::new
+        }
     }
+
+    /// `leak(δ, k, μ, d) := k · [q(δ) + stir(δ)] + T(δ) + ood(δ) + (d+1) · μ`
+    const fn leak(&self) -> usize {
+        self.k * (self.q_delta + self.stir_delta)
+            + self.t_delta
+            + self.ood_delta
+            + (self.d + 1) * self.mu
+    }
+
+    /// Compute `q(δ)` and `stir(δ)` for a given security target and rate.
+    ///
+    /// - `q(δ) = ⌈λ / log₂(1/(1-δ))⌉`
+    /// - `stir(δ) ≈ ⌈λ / (s + log₂(1/(1-δ)))⌉`
+    #[allow(clippy::cast_sign_loss)]
+    fn query_counts(
+        unique_decoding: bool,
+        security_target: f64,
+        rate: f64,
+        folding_factor: usize,
+    ) -> (usize, usize) {
+        let q = irs_commit::num_in_domain_queries(unique_decoding, security_target, rate);
+        let s = folding_factor as f64;
+        let slack = irs_commit::johnson_slack(unique_decoding, rate);
+        let per_sample = if unique_decoding {
+            f64::midpoint(1., rate)
+        } else {
+            rate.sqrt() + slack
+        };
+        let stir = (security_target / (s + (-per_sample.log2()))).ceil() as usize;
+        (q, stir)
+    }
+}
+
+/// Compute `q_ub` — the total leakage upper bound across both WHIR instances.
+///
+/// `q_ub ≤ leak(δ₁, k₁, μ, d) + leak(δ₂, k₂, μ, 3)`
+const fn query_upper_bound(witness: &InstanceLeak, blinding: &InstanceLeak) -> usize {
+    witness.leak() + blinding.leak()
 }
 
 #[cfg(test)]
@@ -243,7 +208,7 @@ mod tests {
 
     use ark_ff::{AdditiveGroup, Field};
 
-    use super::*;
+    use super::Config;
     use crate::{
         algebra::{
             fields::Field64,
@@ -257,20 +222,10 @@ mod tests {
 
     type F = Field64;
 
-    fn make_test_blinding_size_policy() -> BlindingSizePolicy {
-        BlindingSizePolicy {
-            q_delta_1: 4,
-            q_delta_2: 4,
-            t1: 4,
-            t2: 4,
-            sumcheck_round_degree: 3,
-        }
-    }
-
-    const TEST_NUM_VARIABLES: usize = 8;
+    const TEST_NUM_VARIABLES: usize = 12;
     const TEST_NUM_COEFFS: usize = 1 << TEST_NUM_VARIABLES;
 
-    fn make_test_config(num_polynomials: usize) -> Config<F> {
+    fn make_test_config() -> Config<F> {
         let whir_params = ProtocolParameters {
             unique_decoding: false,
             security_level: 16,
@@ -281,21 +236,13 @@ mod tests {
             batch_size: 1,
             hash_id: hash::SHA2,
         };
-        Config::<F>::new_with_blinding_size_policy(
-            1 << TEST_NUM_VARIABLES,
-            &whir_params,
-            num_polynomials,
-            make_test_blinding_size_policy(),
-        )
+        let mut config = Config::new(&whir_params, TEST_NUM_VARIABLES);
+        config.blinded_polynomial.disable_pow();
+        config.blinding_polynomial.disable_pow();
+        config
     }
 
-    fn linear_form_refs(forms: &[Box<dyn LinearForm<F>>]) -> Vec<&dyn LinearForm<F>> {
-        forms
-            .iter()
-            .map(|w| w.as_ref() as &dyn LinearForm<F>)
-            .collect()
-    }
-
+    /// Materialize linear forms into Covectors for the prover.
     fn to_prove_forms(
         forms: &[Box<dyn LinearForm<F>>],
         size: usize,
@@ -305,157 +252,257 @@ mod tests {
             .map(|f| {
                 let mut cv = vec![F::ZERO; size];
                 f.accumulate(&mut cv, F::ONE);
-                Box::new(Covector { vector: cv }) as Box<dyn LinearForm<F>>
+                Box::new(Covector::new(cv)) as Box<dyn LinearForm<F>>
             })
             .collect()
     }
 
-    fn make_two_poly_vectors(mul: u64, add: u64) -> (Vec<F>, Vec<F>) {
-        let v0 = vec![F::ONE; TEST_NUM_COEFFS];
-        let v1 = (0..TEST_NUM_COEFFS)
-            .map(|i| F::from((i as u64).wrapping_mul(mul).wrapping_add(add)))
-            .collect();
-        (v0, v1)
-    }
-
-    fn compute_evaluations(
-        params: &Config<F>,
-        forms: &[&MultilinearExtension<F>],
-        vectors: &[&[F]],
-    ) -> Vec<F> {
-        let embedding = params.blinded_commitment.embedding();
-        let mut evals = Vec::with_capacity(forms.len() * vectors.len());
-        for &form in forms {
-            for &vector in vectors {
-                evals.push(form.evaluate(embedding, vector));
-            }
-        }
-        evals
-    }
-
+    /// Helper: run a full prove → verify cycle for zkWHIR 2.0.
+    /// `vectors` is a list of witness polynomial evaluation tables.
+    /// `evaluations` is row-major: `evaluations[j * n + i]` = ⟨wⱼ, fᵢ⟩.
+    #[allow(clippy::needless_pass_by_value)]
     fn prove_and_verify(
-        params: &Config<F>,
-        vectors: &[&[F]],
-        forms: &[Box<dyn LinearForm<F>>],
+        config: &Config<F>,
+        vectors: Vec<Vec<F>>,
+        forms: Vec<Box<dyn LinearForm<F>>>,
         evaluations: &[F],
-        session_tag: &str,
     ) {
-        let verify_refs = linear_form_refs(forms);
-        let prove_forms = to_prove_forms(forms, params.blinded_commitment.initial_size());
-        let tag = session_tag.to_owned();
-        let ds = DomainSeparator::protocol(params)
-            .session(&tag)
+        let prove_forms = to_prove_forms(forms.as_slice(), vectors[0].len());
+
+        let ds = DomainSeparator::protocol(config)
+            .session(&format!("zk2-pv {}:{}", file!(), line!()))
             .instance(&Empty);
         let mut prover_state = ProverState::new_std(&ds);
-        let witness = params.commit(&mut prover_state, vectors);
-        let _ = params.prove(
+
+        let poly_refs: Vec<&[F]> = vectors.iter().map(|v| v.as_slice()).collect();
+        let witness = config.commit(&mut prover_state, &poly_refs);
+        config.prove(
             &mut prover_state,
-            vectors
-                .iter()
-                .map(|&v| Cow::Borrowed(v))
-                .collect::<Vec<_>>(),
+            vectors.into_iter().map(Cow::Owned).collect(),
             witness,
             prove_forms,
             Cow::Borrowed(evaluations),
         );
+
         let proof = prover_state.proof();
         let mut verifier_state = VerifierState::new_std(&ds, &proof);
-        let commitment = params
-            .receive_commitments(&mut verifier_state, vectors.len())
-            .expect("receive commitments");
-        params
-            .verify(&mut verifier_state, &verify_refs, evaluations, &commitment)
-            .expect("verify zk wrapper");
+
+        let commitments = config
+            .receive_commitments(&mut verifier_state)
+            .expect("receive_commitments failed");
+
+        let weight_refs: Vec<&dyn LinearForm<F>> = forms
+            .iter()
+            .map(|f| f.as_ref() as &dyn LinearForm<F>)
+            .collect();
+
+        config
+            .verify(&mut verifier_state, &weight_refs, evaluations, &commitments)
+            .expect("verification failed");
     }
 
     #[test]
-    fn test_whir_zk_stage1_single_poly() {
+    fn test_zk2_prove_verify_single_point() {
         let mut rng = ark_std::test_rng();
-        let params = make_test_config(1);
+        let config = make_test_config();
 
         let vector = vec![F::ONE; TEST_NUM_COEFFS];
         let point = MultilinearPoint::rand(&mut rng, TEST_NUM_VARIABLES);
         let form = MultilinearExtension { point: point.0 };
-        let evaluation = form.evaluate(params.blinded_commitment.embedding(), &vector);
-        let forms: Vec<Box<dyn LinearForm<F>>> = vec![Box::new(form)];
-        prove_and_verify(
-            &params,
-            &[&vector],
-            &forms,
-            &[evaluation],
-            &format!("zk-stage1 {}:{}", file!(), line!()),
-        );
+        let evaluation = form.evaluate(config.blinded_polynomial.embedding(), &vector);
+
+        prove_and_verify(&config, vec![vector], vec![Box::new(form)], &[evaluation]);
     }
 
     #[test]
-    fn test_whir_zk_stage1_multi_poly_multi_query() {
+    fn test_zk2_prove_verify_multiple_points() {
         let mut rng = ark_std::test_rng();
-        let params = make_test_config(2);
+        let config = make_test_config();
 
-        let (v0, v1) = make_two_poly_vectors(17, 3);
-        let vectors = [&v0[..], &v1[..]];
+        let vector: Vec<F> = (0..TEST_NUM_COEFFS)
+            .map(|i| F::from(i as u64 + 1))
+            .collect();
 
         let p0 = MultilinearPoint::rand(&mut rng, TEST_NUM_VARIABLES);
         let p1 = MultilinearPoint::rand(&mut rng, TEST_NUM_VARIABLES);
         let f0 = MultilinearExtension { point: p0.0 };
         let f1 = MultilinearExtension { point: p1.0 };
-        let evaluations = compute_evaluations(&params, &[&f0, &f1], &vectors);
 
-        let forms: Vec<Box<dyn LinearForm<F>>> = vec![Box::new(f0), Box::new(f1)];
+        let embedding = config.blinded_polynomial.embedding();
+        let eval0 = f0.evaluate(embedding, &vector);
+        let eval1 = f1.evaluate(embedding, &vector);
+
         prove_and_verify(
-            &params,
-            &vectors,
-            &forms,
-            &evaluations,
-            &format!("zk-stage1-multi {}:{}", file!(), line!()),
+            &config,
+            vec![vector],
+            vec![Box::new(f0), Box::new(f1)],
+            &[eval0, eval1],
         );
     }
 
     #[test]
+    fn test_zk2_prove_verify_with_covector() {
+        let mut rng = ark_std::test_rng();
+        let config = make_test_config();
+
+        let vector = vec![F::ONE; TEST_NUM_COEFFS];
+
+        let point = MultilinearPoint::rand(&mut rng, TEST_NUM_VARIABLES);
+        let mle_form = MultilinearExtension { point: point.0 };
+        let embedding = config.blinded_polynomial.embedding();
+        let mle_eval = mle_form.evaluate(embedding, &vector);
+
+        let cov = Covector::new((0..TEST_NUM_COEFFS).map(|i| F::from(i as u64)).collect());
+        let cov_eval = cov.evaluate(embedding, &vector);
+
+        prove_and_verify(
+            &config,
+            vec![vector],
+            vec![Box::new(mle_form), Box::new(cov)],
+            &[mle_eval, cov_eval],
+        );
+    }
+
+    fn make_test_config_batch(batch_size: usize) -> Config<F> {
+        let whir_params = ProtocolParameters {
+            unique_decoding: false,
+            security_level: 16,
+            pow_bits: 0,
+            initial_folding_factor: 2,
+            folding_factor: 2,
+            starting_log_inv_rate: 1,
+            batch_size,
+            hash_id: hash::SHA2,
+        };
+        let mut config = Config::new(&whir_params, TEST_NUM_VARIABLES);
+        config.blinded_polynomial.disable_pow();
+        config.blinding_polynomial.disable_pow();
+        config
+    }
+
+    #[test]
+    fn test_zk2_prove_verify_multi_vector() {
+        let mut rng = ark_std::test_rng();
+        let config = make_test_config_batch(2);
+
+        let v0: Vec<F> = (0..TEST_NUM_COEFFS)
+            .map(|i| F::from(i as u64 + 1))
+            .collect();
+        let v1: Vec<F> = (0..TEST_NUM_COEFFS)
+            .map(|i| F::from(i as u64 * 3 + 7))
+            .collect();
+
+        let p0 = MultilinearPoint::rand(&mut rng, TEST_NUM_VARIABLES);
+        let f0 = MultilinearExtension { point: p0.0 };
+
+        let embedding = config.blinded_polynomial.embedding();
+        // evaluations[j * n + i] = ⟨wⱼ, fᵢ⟩
+        // 1 form, 2 vectors: evaluations = [⟨w₀, f₀⟩, ⟨w₀, f₁⟩]
+        let eval_0_0 = f0.evaluate(embedding, &v0);
+        let eval_0_1 = f0.evaluate(embedding, &v1);
+
+        prove_and_verify(
+            &config,
+            vec![v0, v1],
+            vec![Box::new(f0)],
+            &[eval_0_0, eval_0_1],
+        );
+    }
+
+    #[test]
+    fn test_zk2_prove_verify_multi_vector_multi_form() {
+        let mut rng = ark_std::test_rng();
+        let config = make_test_config_batch(2);
+
+        let v0: Vec<F> = (0..TEST_NUM_COEFFS)
+            .map(|i| F::from(i as u64 + 1))
+            .collect();
+        let v1: Vec<F> = (0..TEST_NUM_COEFFS)
+            .map(|i| F::from(i as u64 * 3 + 7))
+            .collect();
+
+        let p0 = MultilinearPoint::rand(&mut rng, TEST_NUM_VARIABLES);
+        let p1 = MultilinearPoint::rand(&mut rng, TEST_NUM_VARIABLES);
+        let f0 = MultilinearExtension { point: p0.0 };
+        let f1 = MultilinearExtension { point: p1.0 };
+
+        let embedding = config.blinded_polynomial.embedding();
+        // Row-major: evaluations[j * n + i] = ⟨wⱼ, fᵢ⟩
+        // 2 forms × 2 vectors = 4 evaluations
+        let eval_0_0 = f0.evaluate(embedding, &v0);
+        let eval_0_1 = f0.evaluate(embedding, &v1);
+        let eval_1_0 = f1.evaluate(embedding, &v0);
+        let eval_1_1 = f1.evaluate(embedding, &v1);
+
+        prove_and_verify(
+            &config,
+            vec![v0, v1],
+            vec![Box::new(f0), Box::new(f1)],
+            &[eval_0_0, eval_0_1, eval_1_0, eval_1_1],
+        );
+    }
+
+    // =====================================================================
+    // Soundness / negative tests
+    // =====================================================================
+
     /// Verification must reject when the public evaluations are tampered with.
     /// Both `Err` and a panic (from debug transcript checks) count as rejection.
-    fn test_whir_zk_stage1_rejects_wrong_evaluations() {
+    #[test]
+    fn test_zk2_rejects_wrong_evaluations() {
         let mut rng = ark_std::test_rng();
-        let params = make_test_config(2);
+        let config = make_test_config();
 
-        let (v0, v1) = make_two_poly_vectors(29, 7);
-        let vectors = [&v0[..], &v1[..]];
-
+        let vector: Vec<F> = (0..TEST_NUM_COEFFS)
+            .map(|i| F::from(i as u64 + 1))
+            .collect();
         let p0 = MultilinearPoint::rand(&mut rng, TEST_NUM_VARIABLES);
         let p1 = MultilinearPoint::rand(&mut rng, TEST_NUM_VARIABLES);
         let f0 = MultilinearExtension { point: p0.0 };
         let f1 = MultilinearExtension { point: p1.0 };
-        let evaluations = compute_evaluations(&params, &[&f0, &f1], &vectors);
+
+        let embedding = config.blinded_polynomial.embedding();
+        let evaluations = vec![
+            f0.evaluate(embedding, &vector),
+            f1.evaluate(embedding, &vector),
+        ];
 
         let forms: Vec<Box<dyn LinearForm<F>>> = vec![Box::new(f0), Box::new(f1)];
-        let refs = linear_form_refs(&forms);
-        let prove_forms = to_prove_forms(&forms, params.blinded_commitment.initial_size());
-        let ds = DomainSeparator::protocol(&params)
-            .session(&format!("zk-stage1-negative {}:{}", file!(), line!()))
+        let prove_forms = to_prove_forms(&forms, vector.len());
+
+        let ds = DomainSeparator::protocol(&config)
+            .session(&format!("zk2-wrong-eval {}:{}", file!(), line!()))
             .instance(&Empty);
         let mut prover_state = ProverState::new_std(&ds);
-        let witness = params.commit(&mut prover_state, &vectors);
-        let _ = params.prove(
+        let witness = config.commit(&mut prover_state, &[&vector]);
+        config.prove(
             &mut prover_state,
-            vectors
-                .iter()
-                .map(|&v| Cow::Borrowed(v))
-                .collect::<Vec<_>>(),
+            vec![Cow::Owned(vector)],
             witness,
             prove_forms,
             Cow::Borrowed(&evaluations),
         );
-
         let proof = prover_state.proof();
-        let mut wrong_evaluations = evaluations.clone();
+
+        let mut wrong_evaluations = evaluations;
         wrong_evaluations[0] += F::ONE;
+
+        let weight_refs: Vec<&dyn LinearForm<F>> = forms
+            .iter()
+            .map(|f| f.as_ref() as &dyn LinearForm<F>)
+            .collect();
 
         let verify_outcome = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
             let mut verifier_state = VerifierState::new_std(&ds, &proof);
-            let commitment = params
-                .receive_commitments(&mut verifier_state, 2)
-                .expect("receive commitments");
-            params.verify(&mut verifier_state, &refs, &wrong_evaluations, &commitment)
+            let commitments = config
+                .receive_commitments(&mut verifier_state)
+                .expect("receive_commitments");
+            config.verify(
+                &mut verifier_state,
+                &weight_refs,
+                &wrong_evaluations,
+                &commitments,
+            )
         }));
         if let Ok(result) = verify_outcome {
             assert!(
@@ -465,91 +512,106 @@ mod tests {
         }
     }
 
+    /// Verification must reject when the proof transcript is corrupted.
     #[test]
-    fn test_whir_zk_stage1_rejects_tampered_proof() {
+    fn test_zk2_rejects_tampered_proof() {
         let mut rng = ark_std::test_rng();
-        let params = make_test_config(2);
+        let config = make_test_config();
 
-        let (v0, v1) = make_two_poly_vectors(13, 11);
-        let vectors = [&v0[..], &v1[..]];
-
+        let vector: Vec<F> = (0..TEST_NUM_COEFFS)
+            .map(|i| F::from(i as u64 + 1))
+            .collect();
         let p0 = MultilinearPoint::rand(&mut rng, TEST_NUM_VARIABLES);
         let p1 = MultilinearPoint::rand(&mut rng, TEST_NUM_VARIABLES);
         let f0 = MultilinearExtension { point: p0.0 };
         let f1 = MultilinearExtension { point: p1.0 };
-        let evaluations = compute_evaluations(&params, &[&f0, &f1], &vectors);
+
+        let embedding = config.blinded_polynomial.embedding();
+        let evaluations = vec![
+            f0.evaluate(embedding, &vector),
+            f1.evaluate(embedding, &vector),
+        ];
 
         let forms: Vec<Box<dyn LinearForm<F>>> = vec![Box::new(f0), Box::new(f1)];
-        let refs = linear_form_refs(&forms);
-        let prove_forms = to_prove_forms(&forms, params.blinded_commitment.initial_size());
-        let ds = DomainSeparator::protocol(&params)
-            .session(&format!("zk-stage1-tamper {}:{}", file!(), line!()))
+        let prove_forms = to_prove_forms(&forms, vector.len());
+
+        let ds = DomainSeparator::protocol(&config)
+            .session(&format!("zk2-tamper {}:{}", file!(), line!()))
             .instance(&Empty);
         let mut prover_state = ProverState::new_std(&ds);
-        let witness = params.commit(&mut prover_state, &vectors);
-        let _ = params.prove(
+        let witness = config.commit(&mut prover_state, &[&vector]);
+        config.prove(
             &mut prover_state,
-            vectors
-                .iter()
-                .map(|&v| Cow::Borrowed(v))
-                .collect::<Vec<_>>(),
+            vec![Cow::Owned(vector)],
             witness,
             prove_forms,
             Cow::Borrowed(&evaluations),
         );
 
-        let mut proof = prover_state.proof();
-        if let Some(last) = proof.narg_string.last_mut() {
+        let mut tampered_proof = prover_state.proof();
+        if let Some(last) = tampered_proof.narg_string.last_mut() {
             *last ^= 1;
         } else {
             panic!("expected non-empty proof transcript");
         }
 
+        let weight_refs: Vec<&dyn LinearForm<F>> = forms
+            .iter()
+            .map(|f| f.as_ref() as &dyn LinearForm<F>)
+            .collect();
+
         let verify_outcome = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-            let mut verifier_state = VerifierState::new_std(&ds, &proof);
-            let commitment = params
-                .receive_commitments(&mut verifier_state, 2)
-                .expect("receive commitments");
-            params.verify(&mut verifier_state, &refs, &evaluations, &commitment)
+            let mut verifier_state = VerifierState::new_std(&ds, &tampered_proof);
+            let commitments = config
+                .receive_commitments(&mut verifier_state)
+                .expect("receive_commitments");
+            config.verify(
+                &mut verifier_state,
+                &weight_refs,
+                &evaluations,
+                &commitments,
+            )
         }));
         if let Ok(result) = verify_outcome {
             assert!(
                 result.is_err(),
                 "verification should reject tampered proof bytes"
             );
-        } else {
-            // In debug transcript mode, tampering may trigger an interaction-pattern panic
-            // before returning `Err`, which is still a valid rejection.
         }
+        // A panic is also a valid rejection (debug transcript checks).
     }
 
-    /// Soundness exploit: malicious prover generates proof for WRONG evaluation.
-    /// A sound PCS must reject; if verify() returns Ok, g_eval freedom lets
-    /// the prover forge arbitrary evaluation claims.
+    /// Soundness: a malicious prover who generates a proof for a wrong evaluation
+    /// must be rejected. If verify() accepts, it means the prover can forge
+    /// arbitrary evaluation claims.
     #[test]
-    fn test_whir_zk_malicious_prover_wrong_evaluation() {
+    fn test_zk2_malicious_prover_wrong_evaluation() {
         let mut rng = ark_std::test_rng();
-        let params = make_test_config(1);
+        let config = make_test_config();
 
         let vector = vec![F::ONE; TEST_NUM_COEFFS];
         let point = MultilinearPoint::rand(&mut rng, TEST_NUM_VARIABLES);
         let form = MultilinearExtension { point: point.0 };
-        let correct_evaluation = form.evaluate(params.blinded_commitment.embedding(), &vector);
+        let correct_evaluation = form.evaluate(config.blinded_polynomial.embedding(), &vector);
         let wrong_evaluation = correct_evaluation + F::from(42u64);
 
         let forms: Vec<Box<dyn LinearForm<F>>> = vec![Box::new(form)];
-        let refs = linear_form_refs(&forms);
-        let prove_forms = to_prove_forms(&forms, params.blinded_commitment.initial_size());
-        let ds = DomainSeparator::protocol(&params)
-            .session(&format!("zk-malicious {}:{}", file!(), line!()))
+        let prove_forms = to_prove_forms(&forms, vector.len());
+        let weight_refs: Vec<&dyn LinearForm<F>> = forms
+            .iter()
+            .map(|f| f.as_ref() as &dyn LinearForm<F>)
+            .collect();
+
+        let ds = DomainSeparator::protocol(&config)
+            .session(&format!("zk2-malicious {}:{}", file!(), line!()))
             .instance(&Empty);
 
         let outcome = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
             let mut prover_state = ProverState::new_std(&ds);
-            let witness = params.commit(&mut prover_state, &[&vector]);
-            let _ = params.prove(
+            let witness = config.commit(&mut prover_state, &[&vector]);
+            config.prove(
                 &mut prover_state,
-                vec![Cow::Borrowed(&vector)],
+                vec![Cow::Owned(vector.clone())],
                 witness,
                 prove_forms,
                 Cow::Owned(vec![wrong_evaluation]),
@@ -557,17 +619,22 @@ mod tests {
 
             let proof = prover_state.proof();
             let mut verifier_state = VerifierState::new_std(&ds, &proof);
-            let commitment = params
-                .receive_commitments(&mut verifier_state, 1)
-                .expect("receive commitments");
-            params.verify(&mut verifier_state, &refs, &[wrong_evaluation], &commitment)
+            let commitments = config
+                .receive_commitments(&mut verifier_state)
+                .expect("receive_commitments");
+            config.verify(
+                &mut verifier_state,
+                &weight_refs,
+                &[wrong_evaluation],
+                &commitments,
+            )
         }));
 
         if let Ok(result) = outcome {
             assert!(
                 result.is_err(),
                 "SOUNDNESS BUG: verifier accepted wrong evaluation from malicious prover \
-             (correct={correct_evaluation:?}, claimed={wrong_evaluation:?})"
+                 (correct={correct_evaluation:?}, claimed={wrong_evaluation:?})"
             );
         }
     }
