+++ src/protocols/whir_zk/utils.rs
@@ -1,651 +1,763 @@
-use ark_ff::Field;
-use ark_std::rand::{CryptoRng, RngCore};
+use ark_ff::FftField;
 
-#[cfg(not(feature = "parallel"))]
-use crate::algebra::scalar_mul_add;
-use crate::algebra::{
-    embedding::Identity,
-    linear_form::{Covector, Evaluate, LinearForm},
-    univariate_evaluate,
-};
+use super::Config;
+use crate::algebra::{geometric_accumulate, geometric_sequence, MultilinearPoint};
 
-/// Blinding evaluations at one round-0 point `gamma in Gamma`.
+/// Derived protocol dimensions for a single zkWHIR 2.0 execution.
 ///
-/// Values are represented in the doc-faithful `beq((pow(gamma), -rho), .)` basis:
-/// - `m_eval = M(gamma, -rho)`,
-/// - `g_hat_evals[j] = g_hat_j(gamma, -rho)` (with the committed `ell+1` embedding).
-#[derive(Clone, Debug)]
-pub struct BlindingEvaluations<F: Field> {
-    pub gamma: F,
-    pub m_eval: F,
-    pub g_hat_evals: Vec<F>,
+/// Computed once from the config and reused across prover/verifier steps
+/// to avoid recomputing (and passing individually) the same derived values.
+#[derive(Clone, Copy)]
+pub(super) struct ProtocolDims {
+    pub(super) mu: usize,
+    pub(super) ell: usize,
+    pub(super) rem: usize,
+    pub(super) nu: usize,
+    pub(super) size: usize,
+    pub(super) num_vectors: usize,
+    pub(super) num_blinding_vecs: usize,
 }
 
-impl<F: Field> BlindingEvaluations<F> {
-    /// Compute
-    /// `h(gamma) = m(gamma,-rho) + Sum_j beta^j * gamma^(2^(j-1)) * g_hat_j(gamma,-rho)`.
-    pub fn compute_h_value(&self, blinding_challenge: F) -> F {
-        let mut value = self.m_eval;
-        let mut blinding_power = blinding_challenge;
-        let mut gamma_power = self.gamma;
-        for &g_hat_eval in &self.g_hat_evals {
-            value += blinding_power * gamma_power * g_hat_eval;
-            blinding_power *= blinding_challenge;
-            gamma_power = gamma_power.square();
-        }
-        value
-    }
-}
-
-/// Prover-side blinding polynomial family.
-///
-/// Shape follows the protocol notation:
-/// - `m_poly(z, t)` on `ell + 1` variables by interleaving `(g0_hat, msk)`,
-/// - `g_hats[j](z)` on `ell` variables.
-#[derive(Clone, Debug)]
-pub struct BlindingPolynomials<F: Field> {
-    pub m_poly: Vec<F>,
-    pub g_hats: Vec<Vec<F>>,
-}
-
-impl<F: Field> BlindingPolynomials<F> {
-    pub fn sample<R: RngCore + CryptoRng>(
-        rng: &mut R,
-        num_blinding_variables: usize,
-        num_witness_variables: usize,
-    ) -> Self {
-        let half_size = 1usize << num_blinding_variables;
-
-        let msk = (0..half_size).map(|_| F::rand(rng)).collect::<Vec<_>>();
-        let g0_hat = (0..half_size).map(|_| F::rand(rng)).collect::<Vec<_>>();
-        let m_poly: Vec<_> = g0_hat
-            .iter()
-            .zip(&msk)
-            .flat_map(|(&g, &m)| [g, m])
-            .collect();
-        let g_hats = (0..num_witness_variables)
-            .map(|_| (0..half_size).map(|_| F::rand(rng)).collect::<Vec<_>>())
-            .collect::<Vec<_>>();
-
-        Self { m_poly, g_hats }
-    }
-
-    /// Build committed vectors `[M, ĝ_1^emb, ..., ĝ_mu^emb]` on `ell+1` variables.
-    ///
-    /// This matches the blinding WHIR commitment layout used by prover and verifier.
-    pub fn layout_vectors(&self) -> Vec<Vec<F>> {
-        std::iter::once(self.m_poly.clone())
-            .chain(self.embedded_g_hats())
-            .collect()
-    }
-
-    /// Precompute `ell+1` embeddings of all `g_hat` vectors.
-    pub fn embedded_g_hats(&self) -> Vec<Vec<F>> {
-        self.g_hats
-            .iter()
-            .map(|g_hat| embed_to_ell_plus_one::<F>(g_hat))
-            .collect()
-    }
-
-    #[cfg(test)]
-    pub fn evaluate_at(&self, gamma: F, masking_challenge: F) -> BlindingEvaluations<F> {
-        let embedded_g_hats = self.embedded_g_hats();
-        self.evaluate_at_with_embedded(gamma, masking_challenge, &embedded_g_hats)
-    }
-
-    #[cfg(test)]
-    pub fn evaluate_at_with_embedded(
-        &self,
-        gamma: F,
-        masking_challenge: F,
-        embedded_g_hats: &[Vec<F>],
-    ) -> BlindingEvaluations<F> {
-        debug_assert_eq!(self.m_poly.len() % 2, 0, "m_poly length must be even");
-        let half_size = self.m_poly.len() / 2;
-        debug_assert!(
-            self.g_hats.iter().all(|g_hat| g_hat.len() == half_size),
-            "all g_hats must have length {half_size}"
-        );
-        debug_assert_eq!(
-            embedded_g_hats.len(),
-            self.g_hats.len(),
-            "embedded_g_hats count mismatch"
-        );
-        let num_blinding_variables = half_size.ilog2() as usize;
-        let beq_covector = beq_covector_at_gamma(gamma, masking_challenge, num_blinding_variables);
-        self.evaluate_with_covector(&beq_covector, embedded_g_hats, gamma)
-    }
-
-    /// Evaluate blinding polynomials using a precomputed `beq` covector.
-    ///
-    /// Callers that share one `beq_covector` across multiple polynomials at the
-    /// same gamma avoid recomputing the `eq_weights_at_gamma` tensor product for
-    /// each polynomial separately.
-    pub fn evaluate_with_covector(
-        &self,
-        beq_covector: &Covector<F>,
-        embedded_g_hats: &[Vec<F>],
-        gamma: F,
-    ) -> BlindingEvaluations<F> {
-        debug_assert_eq!(embedded_g_hats.len(), self.g_hats.len());
-        let embedding = Identity::<F>::new();
-        let m_eval = beq_covector.evaluate(&embedding, &self.m_poly);
-        let g_hat_evals = embedded_g_hats
-            .iter()
-            .map(|g_hat_embedded| beq_covector.evaluate(&embedding, g_hat_embedded))
-            .collect();
-        BlindingEvaluations {
-            gamma,
-            m_eval,
-            g_hat_evals,
+impl ProtocolDims {
+    pub(super) fn new<F: FftField>(config: &Config<F>, num_vectors: usize) -> Self {
+        let mu = config.blinded_polynomial.initial_num_variables();
+        let ell = config.blinding_polynomial.initial_num_variables() - 1;
+        let rem = mu % ell;
+        let num_blinding_vecs = config.blinding_polynomial.initial_committer.num_vectors;
+        let nu = num_blinding_vecs - num_vectors;
+        let size = 1 << mu;
+        Self {
+            mu,
+            ell,
+            rem,
+            nu,
+            size,
+            num_vectors,
+            num_blinding_vecs,
         }
     }
-}
 
-/// Embed an `ell`-variable coefficient vector into `ell+1` variables by
-/// interleaving zeros: `[c0, c1, ...] -> [c0, 0, c1, 0, ...]`.
-///
-/// This effectively treats the original polynomial as depending only on the
-/// even-indexed variable, placing it in the correct position within the
-/// `(ell+1)`-variable commitment layout used by the blinding WHIR.
-fn embed_to_ell_plus_one<F: Field>(coeffs: &[F]) -> Vec<F> {
-    coeffs.iter().flat_map(|&c| [c, F::ZERO]).collect()
-}
-
-/// Recombine a single combined claim from already tau2-batched
-/// per-vector claims `(m_claim, g_hat_claims)` using tau1-batching.
-pub fn combine_claim_from_components<F: Field>(m_claim: F, g_hat_claims: &[F], tau1: F) -> F {
-    // Factor 2 arises because the combined polynomial is
-    //   C(tau1) = m + 2 * tau1 * Sum_j tau1^j * g_hat_j
-    // The `2` compensates for the embedding interleave: g_hat coefficients
-    // occupy only even slots in the (ell+1)-variable layout, halving the
-    // inner product with the beq covector.
-    const G_HAT_RECOMBINATION_SCALE: u64 = 2;
-    m_claim + F::from(G_HAT_RECOMBINATION_SCALE) * tau1 * univariate_evaluate(g_hat_claims, tau1)
-}
-
-/// Build the combined claims and batched blinding subproof claims from
-/// per-polynomial `(m_claim, g_hat_claims)`.
-///
-/// Returns `(combined_claims, batched_subproof_claims)` where the subproof
-/// layout is `[m_0, g_0_0, ..., g_0_{mu-1}, m_1, g_1_0, ...]`.
-pub fn build_combined_and_subproof_claims<F: Field>(
-    m_claims: &[F],
-    g_hat_claims_per_poly: &[&[F]],
-    tau1: F,
-) -> (Vec<F>, Vec<F>) {
-    let num_polynomials = m_claims.len();
-    debug_assert_eq!(g_hat_claims_per_poly.len(), num_polynomials);
-    let num_witness_variables = g_hat_claims_per_poly.first().map_or(0, |s| s.len());
-    let mut combined_claims = Vec::with_capacity(num_polynomials);
-    let mut subproof_claims = Vec::with_capacity(num_polynomials * (1 + num_witness_variables));
-    for (poly_idx, &m_claim) in m_claims.iter().enumerate() {
-        let g_hat_slice = g_hat_claims_per_poly[poly_idx];
-        subproof_claims.push(m_claim);
-        subproof_claims.extend_from_slice(g_hat_slice);
-        combined_claims.push(combine_claim_from_components(m_claim, g_hat_slice, tau1));
+    /// Number of blinding g-polynomials: ν + 1.
+    pub(super) const fn num_g_polys(&self) -> usize {
+        self.nu + 1
     }
-    (combined_claims, subproof_claims)
 }
 
-/// Build blinding-side linear forms as `[beq_weights, w_folded_weights...]`.
-pub fn build_blinding_forms<'a, F: Field>(
-    beq_weights: &'a Covector<F>,
-    w_folded_weights: &'a [Covector<F>],
-) -> Vec<&'a dyn LinearForm<F>> {
-    std::iter::once(beq_weights as &dyn LinearForm<F>)
-        .chain(w_folded_weights.iter().map(|wf| wf as &dyn LinearForm<F>))
+/// Extract the ℓ-bit sub-index from a μ-bit hypercube index `b` for the Φ_i projection.
+///
+/// Implements the Φ_i morphisms:
+///   Φ₀(x̄) = (x₁, ..., x_ℓ)
+///   Φᵢ(x̄) = (x_{(i-1)·ℓ+rem+1}, ..., x_{i·ℓ+rem})  for i ≥ 1
+///
+/// This is the integer-index (bit-pattern) version: extracts the same ℓ-bit
+/// window that the multivariate Φᵢ would select.
+///
+/// Index convention (big-endian):
+///   index = x_0 · 2^{μ-1} + x_1 · 2^{μ-2} + ... + x_{μ-1} · 2^0
+///
+/// The result is: `(b >> (μ - start - ℓ)) & ((1 << ℓ) - 1)`
+pub(super) const fn phi_i_bits(
+    hypercube_idx: usize,
+    phi_index: usize,
+    mu: usize,
+    ell: usize,
+    rem: usize,
+) -> usize {
+    let start = if phi_index == 0 {
+        0
+    } else {
+        (phi_index - 1) * ell + rem
+    };
+    let shift = mu - start - ell;
+    (hypercube_idx >> shift) & ((1 << ell) - 1)
+}
+
+/// Compute the discrete logarithm of `target` w.r.t. `gen` in a cyclic group
+/// of order `2^log_order`, using the Pohlig-Hellman algorithm.
+///
+/// Returns `i` such that `target == gen^i`, where `0 ≤ i < 2^log_order`.
+/// Panics (in debug builds) if `target` is not in `⟨gen⟩`.
+///
+/// Complexity: O(log_order²) field multiplications — vs O(2^log_order) for linear scan.
+pub(super) fn discrete_log_pow2<F: FftField>(target: F, gen: F, log_order: u32) -> usize {
+    let gen_inv = gen.inverse().expect("generator must be invertible");
+    let mut result = 0usize;
+    let mut current = target;
+    let mut gen_inv_power = gen_inv; // gen^{-2^bit} accumulator
+
+    for bit in 0..log_order {
+        // current^{2^{log_order - bit - 1}} == 1  ⟺  bit `bit` of the index is 0
+        let mut test = current;
+        for _ in 0..(log_order - bit - 1) {
+            test.square_in_place();
+        }
+
+        if test != F::ONE {
+            result |= 1 << bit;
+            current *= gen_inv_power;
+        }
+
+        gen_inv_power.square_in_place();
+    }
+
+    debug_assert_eq!(
+        gen.pow([result as u64]),
+        target,
+        "discrete log verification failed: target not in ⟨gen⟩ of order 2^{log_order}"
+    );
+    result
+}
+
+/// Build the μ-variate evaluation point `fold_args(r̄, z)`.
+///
+/// fold_args(r̄; z) := (r₁, ..., r_s, z^{2⁰}, z^{2¹}, ..., z^{2^{μ-s-1}})
+///
+/// Result: `(r_0, ..., r_{s-1}, z^{2^{k-1}}, z^{2^{k-2}}, ..., z^2, z)`
+/// where `s = |r̄|` and `k = μ − s`.
+///
+/// The z-derived coordinates use descending powers (big-endian convention)
+/// to match the codebase's `UnivariateEvaluation::mle_evaluate` squaring ladder.
+pub(super) fn build_fold_args<F: FftField>(r_bar: &[F], z: F, mu: usize) -> Vec<F> {
+    let num_folded_vars = r_bar.len();
+    let num_z_vars = mu - num_folded_vars;
+    let mut point = Vec::with_capacity(mu);
+    point.extend(r_bar);
+
+    // Squaring ladder: z, z², z⁴, ..., z^{2^{num_z_vars-1}}
+    let mut z_pow = z;
+    let mut z_pows = Vec::with_capacity(num_z_vars);
+    for _ in 0..num_z_vars {
+        z_pows.push(z_pow);
+        z_pow.square_in_place();
+    }
+    // Reverse to descending order: z^{2^{num_z_vars-1}}, ..., z², z
+    point.extend(z_pows.iter().rev());
+    point
+}
+
+/// Build batched eq tables for the blinding proof (Step 7).
+///
+/// Implements the weight polynomial:
+///   wᵢ(z, ȳ) = z · Σⱼ τⱼ · eq(Φᵢ(P[j]), ȳ)
+///
+/// `beq_i[k] = Σ_j τ^{j+1} · Σ_{c,m} eq(r̄, c) · z_j^m · δ(Φ_i(c·M+m), k)`
+#[cfg_attr(feature = "tracing", tracing::instrument(skip_all, fields(num_points = lambda_z_points.len(), mu = dims.mu, ell = dims.ell, num_g_polys = dims.num_g_polys())))]
+pub(super) fn build_beq_tables<F: FftField>(
+    lambda_z_points: &[F],
+    eq_weights: &[F],
+    tau: F,
+    dims: ProtocolDims,
+) -> Vec<Vec<F>> {
+    let mu = dims.mu;
+    let ell = dims.ell;
+    let rem = dims.rem;
+    let num_g_polys = dims.num_g_polys();
+    let half_size = 1usize << ell;
+    let num_folding_vars = eq_weights.len().trailing_zeros() as usize;
+    assert!(
+        num_folding_vars <= ell,
+        "folding factor num_folding_vars={num_folding_vars} must not exceed ell={ell} (would underflow m_cap in Φ₀ window)"
+    );
+    let num_m_bits = mu - num_folding_vars; // number of m-bits (log2 of sub-polynomial length M)
+
+    // Precompute τ powers: [τ, τ², ..., τ^num_points]
+    let tau_powers_full = geometric_sequence(tau, lambda_z_points.len() + 1);
+    let tau_powers = &tau_powers_full[1..];
+
+    // Precompute squaring ladders z^{2^0}, z^{2^1}, ..., z^{2^{num_m_bits-1}} for each z-point
+    let z_pows_all: Vec<Vec<F>> = lambda_z_points
+        .iter()
+        .map(|z| {
+            let mut z_pows = Vec::with_capacity(num_m_bits);
+            let mut z_pow = *z;
+            for _ in 0..num_m_bits {
+                z_pows.push(z_pow);
+                z_pow.square_in_place();
+            }
+            z_pows
+        })
+        .collect();
+
+    let num_points = lambda_z_points.len();
+    let mut tables = vec![vec![F::ZERO; half_size]; num_g_polys];
+
+    for (i, table) in tables.iter_mut().enumerate() {
+        let start_i = if i == 0 { 0 } else { (i - 1) * ell + rem };
+
+        // Bit-window decomposition relative to c|m boundary at position num_folding_vars
+        let a_below = mu - start_i - ell; // free m-bits below window (= shift_i)
+        let a_above = start_i.saturating_sub(num_folding_vars); // free m-bits above
+        let m_cap = num_m_bits - a_below - a_above; // captured m-bits in window
+        let c_cap = ell - m_cap; // captured c-bits in window (low c_cap bits of c)
+
+        // Partial eq marginalization: eq_partial[k_c] = Σ_{c: c & mask = k_c} eq[c]
+        let eq_partial = if c_cap > 0 {
+            let mut eq_partial = vec![F::ZERO; 1 << c_cap];
+            let c_mask = (1 << c_cap) - 1;
+            for (c_idx, &weight) in eq_weights.iter().enumerate() {
+                eq_partial[c_idx & c_mask] += weight;
+            }
+            eq_partial
+        } else {
+            vec![F::ONE] // Σ eq_weights = 1
+        };
+
+        // Build scalars w_j = tp_j · geo_below_j · geo_above_j
+        // and bases base_j = z_j^{2^{a_below}} for geometric_accumulate
+        let m_cap_size = 1usize << m_cap;
+        let mut scalars = Vec::with_capacity(num_points);
+        let mut bases = Vec::with_capacity(num_points);
+
+        for (j, &tp) in tau_powers.iter().enumerate() {
+            let z_pows = &z_pows_all[j];
+
+            // Free-bit product below: Π_{j=0}^{a_below-1} (1 + z^{2^j})
+            let mut geo_below = F::ONE;
+            for &zp in z_pows.iter().take(a_below) {
+                geo_below *= F::ONE + zp;
+            }
+
+            // Free-bit product above: Π_{j=0}^{a_above-1} (1 + z^{2^{a_below+m_cap+j}})
+            let mut geo_above = F::ONE;
+            for &zp in z_pows.iter().skip(a_below + m_cap).take(a_above) {
+                geo_above *= F::ONE + zp;
+            }
+
+            scalars.push(tp * geo_below * geo_above);
+            bases.push(if a_below < num_m_bits {
+                z_pows[a_below]
+            } else {
+                F::ONE
+            });
+        }
+
+        // m_inner[k_m] = Σ_j w_j · base_j^{k_m}
+        let mut m_inner = vec![F::ZERO; m_cap_size];
+        geometric_accumulate(&mut m_inner, scalars, &bases);
+
+        // Assemble: tables[i][k_c · 2^m_cap + k_m] = eq_partial[k_c] · m_inner[k_m]
+        if c_cap > 0 {
+            for (k_c, &ep) in eq_partial.iter().enumerate() {
+                for (k_m, &mi) in m_inner.iter().enumerate() {
+                    table[k_c * m_cap_size + k_m] = ep * mi;
+                }
+            }
+        } else {
+            *table = m_inner;
+        }
+    }
+
+    tables
+}
+
+/// Precompute RS-fold coefficient vectors for the blinding polynomials (Steps 5-6).
+///
+/// Used to evaluate m̃(r̄, z, ρ) and g̃ᵢ(r̄, z) at OOD/STIR/Γ points.
+///
+/// After the initial sumcheck folds `s` variables with randomness `r̄`, the original
+/// 2^μ-coefficient polynomial is viewed as `k = 2^s` sub-polynomials of length `M = 2^(μ-s)`.
+/// The RS-fold is:
+///
+/// ```text
+/// fold(r̄, poly)(z) = Σ_m [ Σ_j eq(r̄, j) · poly[j·M + m] ] · z^m
+/// ```
+///
+/// For blinding polynomials, the 2^μ evaluation table is the lift of an ℓ-variate table
+/// via Φ_i projections: `lifted[b] = table[Φ_i_bits(b)]`.
+///
+/// With multi-polynomial batching (n witness polynomials, batching coefficients α):
+/// - `m_coeffs_all[0][m] = Σ_j eq · [ĝ₀[Φ₀(j·M+m)] + (-ρ)·msk₀[Φ₀(j·M+m)]]`
+/// - `m_coeffs_all[i][m] = (-ρ·αⁱ) · Σ_j eq · mskᵢ[Φ₀(j·M+m)]`  for i = 1..n-1
+/// - `g_i_coeffs[i][m] = Σ_j eq · ĝ_{i+1}[Φ_{i+1}(j·M+m)]`  for i = 0..ν-1
+///
+/// Returns `(m_coeffs_all, g_i_coeffs)` where each vector has length M.
+#[cfg_attr(feature = "tracing", tracing::instrument(skip_all, fields(mu = dims.mu, ell = dims.ell, num_g_polys = g_polys.len())))]
+pub(super) fn compute_rs_fold_blinding_coeffs<F: FftField>(
+    eq_weights: &[F],
+    g_polys: &[Vec<F>],
+    masking_polys: &[Vec<F>],
+    alpha_coeffs: &[F],
+    rho: F,
+    dims: ProtocolDims,
+) -> (Vec<Vec<F>>, Vec<Vec<F>>) {
+    let mu = dims.mu;
+    let ell = dims.ell;
+    let rem = dims.rem;
+    let num_folding_vars = eq_weights.len().trailing_zeros() as usize;
+    let num_sub_polys = 1usize << num_folding_vars;
+    let sub_poly_len = 1usize << (mu - num_folding_vars);
+    let num_g_polys = g_polys.len();
+    let num_masking = masking_polys.len();
+    let neg_rho = -rho;
+
+    // Accumulate g₀ fold coeffs, per-masking fold coeffs, and g_i fold coeffs
+    let accumulate_j =
+        |g0_acc: &mut Vec<F>, msk_accs: &mut Vec<Vec<F>>, g_acc: &mut Vec<Vec<F>>, j: usize| {
+            let eq_j = eq_weights[j];
+            for sub_idx in 0..sub_poly_len {
+                let full_idx = j * sub_poly_len + sub_idx;
+                let phi_0_idx = phi_i_bits(full_idx, 0, mu, ell, rem);
+
+                g0_acc[sub_idx] += eq_j * g_polys[0][phi_0_idx];
+                for (i, msk) in masking_polys.iter().enumerate() {
+                    msk_accs[i][sub_idx] += eq_j * msk[phi_0_idx];
+                }
+
+                for gi in 1..num_g_polys {
+                    let phi_i_idx = phi_i_bits(full_idx, gi, mu, ell, rem);
+                    g_acc[gi - 1][sub_idx] += eq_j * g_polys[gi][phi_i_idx];
+                }
+            }
+        };
+
+    // Assemble m_coeffs_all from raw g₀ fold and masking folds
+    let assemble = |g0_fold: Vec<F>,
+                    msk_folds: Vec<Vec<F>>,
+                    g_i_coeffs: Vec<Vec<F>>|
+     -> (Vec<Vec<F>>, Vec<Vec<F>>) {
+        let mut m_coeffs_all = Vec::with_capacity(num_masking);
+        // m₀ = g₀_fold + (-ρ) · msk₀_fold
+        let m0: Vec<F> = g0_fold
+            .iter()
+            .zip(msk_folds[0].iter())
+            .map(|(&g, &msk)| g + neg_rho * msk)
+            .collect();
+        m_coeffs_all.push(m0);
+        // mᵢ = (-ρ·αⁱ) · mskᵢ_fold for i ≥ 1
+        for i in 1..num_masking {
+            let scale = neg_rho * alpha_coeffs[i];
+            let mi: Vec<F> = msk_folds[i].iter().map(|&v| scale * v).collect();
+            m_coeffs_all.push(mi);
+        }
+        (m_coeffs_all, g_i_coeffs)
+    };
+
+    let mut g0_fold = vec![F::ZERO; sub_poly_len];
+    let mut msk_folds = vec![vec![F::ZERO; sub_poly_len]; num_masking];
+    let mut g_i_coeffs = vec![vec![F::ZERO; sub_poly_len]; num_g_polys - 1];
+    for j in 0..num_sub_polys {
+        accumulate_j(&mut g0_fold, &mut msk_folds, &mut g_i_coeffs, j);
+    }
+
+    assemble(g0_fold, msk_folds, g_i_coeffs)
+}
+
+/// Build weight covectors for Step 7's batched blinding proof.
+///
+/// Constructs `n + ν` covectors used identically by both prover and verifier:
+///   - `w_0`:      `beq_0[k]` for g₀, `(-ρ)·beq_0[k]` for msk₀
+///   - `w_i`:      `(-ρ·αⁱ)·beq_0[k]` for mskᵢ  (1 ≤ i < num_vectors)
+///   - `w_{n+j}`:  `beq_{j+1}[k]` for ĝ_{j+1}    (0 ≤ j < ν)
+pub(super) fn build_weight_covectors<F: FftField>(
+    beq_tables: &[Vec<F>],
+    rho: F,
+    alpha_coeffs: &[F],
+    dims: ProtocolDims,
+) -> Vec<Vec<F>> {
+    let num_vectors = dims.num_vectors;
+    let num_blinding_vecs = dims.num_blinding_vecs;
+    let half_size = 1usize << dims.ell;
+    let full_size = 1usize << (dims.ell + 1);
+
+    let mut weight_covectors: Vec<Vec<F>> = Vec::with_capacity(num_blinding_vecs);
+
+    // w_0: first M-polynomial weight (includes g₀ and msk₀)
+    {
+        let mut w0 = vec![F::ZERO; full_size];
+        let neg_rho = -rho;
+        for k in 0..half_size {
+            w0[2 * k] = beq_tables[0][k];
+            w0[2 * k + 1] = neg_rho * beq_tables[0][k];
+        }
+        weight_covectors.push(w0);
+    }
+
+    // w_i (1 ≤ i < n): additional M-polynomial weights (masking only, no g₀)
+    for &alpha in &alpha_coeffs[1..num_vectors] {
+        let mut wi = vec![F::ZERO; full_size];
+        let scale = -rho * alpha;
+        for k in 0..half_size {
+            wi[2 * k + 1] = scale * beq_tables[0][k];
+        }
+        weight_covectors.push(wi);
+    }
+
+    // w_{n+j-1} (1 ≤ j ≤ ν): ĝ_j weights
+    for beq_table in beq_tables.iter().skip(1) {
+        let mut wj = vec![F::ZERO; full_size];
+        for k in 0..half_size {
+            wj[2 * k] = beq_table[k];
+        }
+        weight_covectors.push(wj);
+    }
+
+    weight_covectors
+}
+
+/// Map gamma points (elements of Ω₁) to their corresponding indices in the
+/// initial codeword [[f̂]].
+///
+/// Each γ ∈ Ω₁ is a power of the round-0 generator. The discrete log gives
+/// the index within the round-0 domain, and multiplying by `stride = |Ω₀|/|Ω₁|`
+/// recovers the position in the initial codeword.
+///
+/// Used identically by both prover (to open [[f̂]]) and verifier (to verify openings).
+pub(super) fn gamma_to_f_hat_indices<F: FftField>(
+    gamma_points: &[F],
+    config: &super::Config<F>,
+) -> Vec<usize> {
+    let initial_codeword_len = config.blinded_polynomial.initial_committer.codeword_length;
+    let round0_codeword_len = config.blinded_polynomial.round_configs[0]
+        .irs_committer
+        .codeword_length;
+    let stride = initial_codeword_len / round0_codeword_len;
+    let gen_h = config.blinded_polynomial.round_configs[0]
+        .irs_committer
+        .generator();
+    let log_round0_len = round0_codeword_len.trailing_zeros();
+
+    gamma_points
+        .iter()
+        .map(|&gamma| discrete_log_pow2(gamma, gen_h, log_round0_len) * stride)
         .collect()
 }
 
-/// Fold a full-size weight to the masking period `2^(ℓ+1)`.
-///
-/// For a weight `w` on `μ` variables and period `P = 2^(ℓ+1)`:
-///   `w_folded[j] = Σ_{i ≡ j (mod P)} w[i]`
-///
-/// The inner product `⟨w_folded, m_poly⟩` equals `⟨w, periodic_extension(m_poly)⟩`,
-/// i.e. the masking contribution `M_eval` used for evaluation binding.
-#[cfg_attr(
-    feature = "tracing",
-    tracing::instrument(skip_all, name = "fold_weight_to_mask_size")
-)]
-pub fn fold_weight_to_mask_size<F: Field>(
-    weight: &dyn crate::algebra::linear_form::LinearForm<F>,
-    num_witness_variables: usize,
-    num_blinding_variables: usize,
-) -> Covector<F> {
-    use std::any::Any;
-
-    let mask_size = 1usize << (num_blinding_variables + 1);
-
-    // Fast path: borrow the vector directly if the weight is already a Covector.
-    let borrowed = (weight as &dyn Any).downcast_ref::<Covector<F>>();
-    let owned;
-    let vector = if let Some(cov) = borrowed {
-        &cov.vector
-    } else {
-        owned = Covector::from(weight);
-        &owned.vector
-    };
-    debug_assert_eq!(vector.len(), 1usize << num_witness_variables);
-
-    fold_vector_to_mask_size(vector, mask_size)
+/// Compute eq_weights from r_bar. Shared helper to avoid redundant computation.
+pub(super) fn compute_eq_weights<F: FftField>(r_bar: &[F]) -> Vec<F> {
+    MultilinearPoint(r_bar.to_vec()).eq_weights()
 }
 
-fn fold_vector_to_mask_size<F: Field>(vector: &[F], mask_size: usize) -> Covector<F> {
-    #[cfg(feature = "parallel")]
-    let folded = {
-        use rayon::prelude::*;
-        vector
-            .par_chunks(mask_size)
-            .fold(
-                || vec![F::ZERO; mask_size],
-                |mut acc, chunk| {
-                    for (j, &v) in chunk.iter().enumerate() {
-                        acc[j] += v;
-                    }
-                    acc
-                },
-            )
-            .reduce(
-                || vec![F::ZERO; mask_size],
-                |mut a, b| {
-                    for (j, &v) in b.iter().enumerate() {
-                        a[j] += v;
-                    }
-                    a
-                },
-            )
-    };
+/// Accumulator for blinding polynomial claims across OOD, STIR, and Γ queries.
+///
+/// Collects (z, m_evals, g_evals) tuples during Steps 5-6 for use in Step 7.
+pub(super) struct LambdaAccumulator<F> {
+    pub(super) z_points: Vec<F>,
+    pub(super) m_evals: Vec<Vec<F>>,
+    pub(super) g_evals: Vec<Vec<F>>,
+}
 
-    #[cfg(not(feature = "parallel"))]
-    let folded = {
-        let mut f = vec![F::ZERO; mask_size];
-        for (i, &v) in vector.iter().enumerate() {
-            f[i % mask_size] += v;
+impl<F> LambdaAccumulator<F> {
+    pub(super) const fn new() -> Self {
+        Self {
+            z_points: Vec::new(),
+            m_evals: Vec::new(),
+            g_evals: Vec::new(),
         }
-        f
-    };
-
-    Covector::new(folded)
-}
-
-/// Build the single batched beq linear form used by the blinding subproof:
-/// `Sum_i tau2^i * beq((pow(gamma_i), -rho), .)`.
-#[cfg_attr(
-    feature = "tracing",
-    tracing::instrument(skip_all, name = "construct_batched_eq_weights")
-)]
-pub fn construct_batched_eq_weights_from_gammas<F: Field>(
-    gammas: &[F],
-    masking_challenge: F,
-    tau2: F,
-    num_blinding_variables: usize,
-) -> Covector<F> {
-    let weight_size = 1 << (num_blinding_variables + 1);
-
-    #[cfg(feature = "parallel")]
-    {
-        use rayon::prelude::*;
-
-        // Pre-compute tau2 powers: tau2^0, tau2^1, ..., tau2^(n-1)
-        let tau2_powers: Vec<F> = {
-            let mut powers = Vec::with_capacity(gammas.len());
-            let mut p = F::ONE;
-            for _ in 0..gammas.len() {
-                powers.push(p);
-                p *= tau2;
-            }
-            powers
-        };
-
-        // Each gamma's eq_weights computation is independent — parallelise across gammas
-        // and reduce by element-wise summation.
-        // The (acc, buf) tuple lets each rayon task reuse one scratch buffer across its gammas.
-        let (weight_evals, _) = gammas
-            .par_iter()
-            .zip(tau2_powers.par_iter())
-            .fold(
-                || (vec![F::ZERO; weight_size], vec![F::ZERO; weight_size]),
-                |(mut acc, mut buf), (&gamma, &power)| {
-                    fill_eq_weights_at_gamma(
-                        &mut buf,
-                        gamma,
-                        masking_challenge,
-                        num_blinding_variables,
-                    );
-                    for (a, &b) in acc.iter_mut().zip(buf.iter()) {
-                        *a += power * b;
-                    }
-                    (acc, buf)
-                },
-            )
-            .reduce(
-                || (vec![F::ZERO; weight_size], vec![F::ZERO; weight_size]),
-                |(mut a, buf_a), (b, _)| {
-                    for (ai, &bi) in a.iter_mut().zip(b.iter()) {
-                        *ai += bi;
-                    }
-                    (a, buf_a)
-                },
-            );
-
-        Covector::new(weight_evals)
     }
 
-    #[cfg(not(feature = "parallel"))]
-    {
-        let mut weight_evals = vec![F::ZERO; weight_size];
-        let mut buf = vec![F::ZERO; weight_size];
-        let mut batching_power = F::ONE;
-        for &gamma in gammas {
-            fill_eq_weights_at_gamma(&mut buf, gamma, masking_challenge, num_blinding_variables);
-            scalar_mul_add(&mut weight_evals, batching_power, &buf);
-            batching_power *= tau2;
-        }
-        Covector::new(weight_evals)
-    }
-}
-
-/// Build `beq((pow(gamma), -rho), .)` as a covector over `ell+1` variables.
-#[cfg(test)]
-pub fn beq_covector_at_gamma<F: Field>(
-    gamma: F,
-    masking_challenge: F,
-    num_blinding_variables: usize,
-) -> Covector<F> {
-    Covector::new(eq_weights_at_gamma(
-        gamma,
-        masking_challenge,
-        num_blinding_variables,
-    ))
-}
-
-#[cfg(test)]
-fn eq_weights_at_gamma<F: Field>(
-    gamma: F,
-    masking_challenge: F,
-    num_blinding_variables: usize,
-) -> Vec<F> {
-    let num_total_variables = num_blinding_variables + 1;
-    let size = 1usize << num_total_variables;
-    let mut weights = vec![F::ZERO; size];
-    fill_eq_weights_at_gamma(
-        &mut weights,
-        gamma,
-        masking_challenge,
-        num_blinding_variables,
-    );
-    weights
-}
-
-/// In-place version of [`eq_weights_at_gamma`] that fills an existing buffer.
-///
-/// `buf` must have length `>= 2^(num_blinding_variables + 1)`.  Reusing a
-/// single buffer across gamma iterations avoids one 8 KiB allocation per point.
-pub fn fill_eq_weights_at_gamma<F: Field>(
-    buf: &mut [F],
-    gamma: F,
-    masking_challenge: F,
-    num_blinding_variables: usize,
-) {
-    // beq((pow(gamma), -rho), .) on ell+1 variables via iterative tensor expansion.
-    // O(2^(ell+1)) muls instead of O(ell * 2^ell) from per-point eq_poly + tensor_product.
-    let num_total_variables = num_blinding_variables + 1; // +1 for the masking variable
-    let size = 1usize << num_total_variables;
-    debug_assert!(buf.len() >= size);
-
-    for w in &mut buf[..size] {
-        *w = F::ZERO;
-    }
-    buf[0] = F::ONE;
-
-    // Variables 0..ell: squaring-ladder basis (gamma, gamma^2, gamma^4, ...)
-    let mut gamma_power = gamma;
-    for i in 0..num_blinding_variables {
-        for j in (0..1usize << i).rev() {
-            buf[2 * j + 1] = buf[j] * gamma_power;
-            buf[2 * j] = buf[j] - buf[2 * j + 1];
-        }
-        gamma_power = gamma_power.square();
+    pub(super) fn push(&mut self, z: F, m: Vec<F>, g: Vec<F>) {
+        debug_assert!(
+            self.m_evals.is_empty() || m.len() == self.m_evals[0].len(),
+            "m_evals length mismatch: expected {}, got {}",
+            self.m_evals.first().map_or(0, Vec::len),
+            m.len()
+        );
+        debug_assert!(
+            self.g_evals.is_empty() || g.len() == self.g_evals[0].len(),
+            "g_evals length mismatch: expected {}, got {}",
+            self.g_evals.first().map_or(0, Vec::len),
+            g.len()
+        );
+        self.z_points.push(z);
+        self.m_evals.push(m);
+        self.g_evals.push(g);
     }
 
-    // Last variable (ell): fixed to -masking_challenge.
-    let neg_rho = -masking_challenge;
-    let half = 1usize << num_blinding_variables;
-    for j in (0..half).rev() {
-        buf[2 * j + 1] = buf[j] * neg_rho;
-        buf[2 * j] = buf[j] - buf[2 * j + 1];
-    }
-}
-
-/// Half-size beq weights: only the first `ell` squaring-ladder variables,
-/// WITHOUT the final `-rho` expansion.
-///
-/// `buf` must have length `>= 2^num_blinding_variables`.
-///
-/// The full-size beq satisfies `beq_full[2j] = beq_half[j] * (1+rho)` and
-/// `beq_full[2j+1] = beq_half[j] * (-rho)`, so any inner product with a
-/// vector whose odd entries are zero reduces to a half-size mixed dot
-/// multiplied by `(1+rho)`.
-pub fn fill_eq_weights_at_gamma_half<F: Field>(
-    buf: &mut [F],
-    gamma: F,
-    num_blinding_variables: usize,
-) {
-    let size = 1usize << num_blinding_variables;
-    debug_assert!(buf.len() >= size);
-
-    for w in &mut buf[..size] {
-        *w = F::ZERO;
-    }
-    buf[0] = F::ONE;
-
-    let mut gamma_power = gamma;
-    for i in 0..num_blinding_variables {
-        for j in (0..1usize << i).rev() {
-            buf[2 * j + 1] = buf[j] * gamma_power;
-            buf[2 * j] = buf[j] - buf[2 * j + 1];
-        }
-        gamma_power = gamma_power.square();
+    #[must_use]
+    pub(super) const fn len(&self) -> usize {
+        self.z_points.len()
     }
 }
 
 #[cfg(test)]
 mod tests {
-    use ark_ff::{AdditiveGroup, Field};
+    use ark_ff::{FftField, Field};
+    use proptest::prelude::*;
 
-    use super::*;
-    use crate::algebra::{embedding::Identity, fields::Field64};
+    use super::{discrete_log_pow2, phi_i_bits};
+    use crate::algebra::fields::Field64;
 
-    fn batch_with_challenge<F: Field, I: IntoIterator<Item = F>>(values: I, challenge: F) -> F {
-        let mut acc = F::ZERO;
-        let mut power = F::ONE;
-        for value in values {
-            acc += power * value;
-            power *= challenge;
-        }
-        acc
-    }
-
-    fn compute_per_polynomial_claims<F: Field>(
-        blinding_evals: &[BlindingEvaluations<F>],
-        tau2: F,
-    ) -> (F, Vec<F>) {
-        let num_g_hats = blinding_evals
-            .first()
-            .map_or(0, |eval| eval.g_hat_evals.len());
-        let m_claim = batch_with_challenge(blinding_evals.iter().map(|eval| eval.m_eval), tau2);
-        let mut g_hat_claims = vec![F::ZERO; num_g_hats];
-        for (g_hat_idx, claim) in g_hat_claims.iter_mut().enumerate() {
-            *claim = batch_with_challenge(
-                blinding_evals
-                    .iter()
-                    .map(|eval| eval.g_hat_evals[g_hat_idx]),
-                tau2,
-            );
-        }
-        (m_claim, g_hat_claims)
-    }
-
-    /// Convenience wrapper for tests that have already materialized [`BlindingEvaluations`].
-    fn construct_batched_eq_weights<F: Field>(
-        blinding_evals: &[BlindingEvaluations<F>],
-        masking_challenge: F,
-        tau2: F,
-        num_blinding_variables: usize,
-    ) -> Covector<F> {
-        let gammas = blinding_evals.iter().map(|e| e.gamma).collect::<Vec<_>>();
-        construct_batched_eq_weights_from_gammas(
-            &gammas,
-            masking_challenge,
-            tau2,
-            num_blinding_variables,
-        )
-    }
-
-    type F = Field64;
+    // ---------------------------------------------------------------
+    // phi_i_bits unit tests
+    // ---------------------------------------------------------------
 
     #[test]
-    fn test_compute_h_value_uses_square_ladder() {
-        let gamma = F::from(5u64);
-        let beta = F::from(7u64);
-        let eval = BlindingEvaluations {
-            gamma,
-            m_eval: F::from(11u64),
-            g_hat_evals: vec![F::from(13u64), F::from(17u64), F::from(19u64)],
-        };
-        let expected = eval.m_eval
-            + beta * gamma * eval.g_hat_evals[0]
-            + beta.square() * gamma.square() * eval.g_hat_evals[1]
-            + beta.pow([3]) * gamma.pow([4]) * eval.g_hat_evals[2];
-        assert_eq!(eval.compute_h_value(beta), expected);
+    fn phi_i_bits_phi0_extracts_top_ell_bits() {
+        // mu=8, ell=3, rem=2 (8 % 3 = 2)
+        // Φ₀ extracts bits [0..3) = top 3 bits
+        let mu = 8;
+        let ell = 3;
+        let rem = mu % ell; // 2
+
+        // b = 0b_101_01010 = 0xAA = 170
+        // top 3 bits: 101 = 5
+        assert_eq!(phi_i_bits(0b1010_1010, 0, mu, ell, rem), 0b101);
+        // b = 0b_111_00000 = 224
+        assert_eq!(phi_i_bits(0b1110_0000, 0, mu, ell, rem), 0b111);
+        // b = 0b_000_11111 = 31
+        assert_eq!(phi_i_bits(0b0001_1111, 0, mu, ell, rem), 0b000);
     }
 
     #[test]
-    fn test_evaluate_at_matches_beq_layout_openings() {
-        let num_blinding_variables = 3usize;
-        let num_witness_variables = 2usize;
-        let half = 1usize << num_blinding_variables;
-        let msk = (0..half)
-            .map(|i| F::from((i as u64) + 2))
-            .collect::<Vec<_>>();
-        let g0_hat = (0..half)
-            .map(|i| F::from((3 * i as u64) + 1))
-            .collect::<Vec<_>>();
-        let mut m_poly = vec![F::from(0u64); 2 * half];
-        for i in 0..half {
-            m_poly[2 * i] = g0_hat[i];
-            m_poly[2 * i + 1] = msk[i];
+    fn phi_i_bits_phi1_extracts_after_rem() {
+        // mu=8, ell=3, rem=2
+        // Φ₁: start = (1-1)*3 + 2 = 2, extracts bits [2..5)
+        let mu = 8;
+        let ell = 3;
+        let rem = 2;
+
+        // b = 0b_10_110_010 = 178
+        // bits [2..5) = 110 = 6
+        assert_eq!(phi_i_bits(0b1011_0010, 1, mu, ell, rem), 0b110);
+    }
+
+    #[test]
+    fn phi_i_bits_phi2_extracts_next_window() {
+        // mu=8, ell=3, rem=2
+        // Φ₂: start = (2-1)*3 + 2 = 5, extracts bits [5..8)
+        let mu = 8;
+        let ell = 3;
+        let rem = 2;
+
+        // b = 0b_10110_101 = 181
+        // bits [5..8) = 101 = 5
+        assert_eq!(phi_i_bits(0b1011_0101, 2, mu, ell, rem), 0b101);
+    }
+
+    #[test]
+    fn phi_i_bits_rem_zero() {
+        // mu=6, ell=3, rem=0 (6 % 3 = 0)
+        // Φ₀: start=0, extracts bits [0..3) = top 3
+        // Φ₁: start=0+0=0, wait: (1-1)*3+0 = 0 ... same as Φ₀
+        // Actually with rem=0, Φ₁ start = (1-1)*3 + 0 = 0, same window.
+        // Let's use mu=6, ell=2, rem=0
+        let mu = 6;
+        let ell = 2;
+        let rem = 0; // 6 % 2 = 0
+
+        // Φ₀: start=0, bits [0..2) = top 2
+        // Φ₁: start=(1-1)*2+0=0, bits [0..2) = same as Φ₀
+        // Φ₂: start=(2-1)*2+0=2, bits [2..4)
+        // Φ₃: start=(3-1)*2+0=4, bits [4..6)
+
+        // b = 0b_11_00_10 = 50
+        assert_eq!(phi_i_bits(0b11_00_10, 0, mu, ell, rem), 0b11);
+        assert_eq!(phi_i_bits(0b11_00_10, 2, mu, ell, rem), 0b00);
+        assert_eq!(phi_i_bits(0b11_00_10, 3, mu, ell, rem), 0b10);
+    }
+
+    #[test]
+    fn phi_i_bits_single_bit_ell() {
+        // ell=1: each Φ extracts a single bit
+        // mu=4, ell=1, rem=0
+        // b = 0b_1011, big-endian positions: [1, 0, 1, 1]
+        //
+        // Φ₀: start=0                → bit position 0 (MSB) = 1
+        // Φ₁: start=(0)*1+0=0        → bit position 0 (same as Φ₀) = 1
+        // Φ₂: start=(1)*1+0=1        → bit position 1 = 0
+        // Φ₃: start=(2)*1+0=2        → bit position 2 = 1
+        // Φ₄: start=(3)*1+0=3        → bit position 3 (LSB) = 1
+        let mu = 4;
+        let ell = 1;
+        let rem = 0;
+
+        assert_eq!(phi_i_bits(0b1011, 0, mu, ell, rem), 1);
+        assert_eq!(phi_i_bits(0b1011, 1, mu, ell, rem), 1); // overlaps with Φ₀ when rem=0
+        assert_eq!(phi_i_bits(0b1011, 2, mu, ell, rem), 0);
+        assert_eq!(phi_i_bits(0b1011, 3, mu, ell, rem), 1);
+        assert_eq!(phi_i_bits(0b1011, 4, mu, ell, rem), 1);
+    }
+
+    #[test]
+    fn phi_i_bits_boundary_all_ones() {
+        let mu = 6;
+        let ell = 3;
+        let rem = 0;
+
+        // b = 0b_111_111 = 63
+        assert_eq!(phi_i_bits(0b111_111, 0, mu, ell, rem), 0b111);
+        assert_eq!(phi_i_bits(0b111_111, 2, mu, ell, rem), 0b111);
+    }
+
+    #[test]
+    fn phi_i_bits_boundary_all_zeros() {
+        let mu = 6;
+        let ell = 3;
+        let rem = 0;
+
+        assert_eq!(phi_i_bits(0, 0, mu, ell, rem), 0);
+        assert_eq!(phi_i_bits(0, 2, mu, ell, rem), 0);
+    }
+
+    // ---------------------------------------------------------------
+    // discrete_log_pow2 unit tests
+    // ---------------------------------------------------------------
+
+    /// Get a generator of the multiplicative subgroup of order 2^k in Field64.
+    fn subgroup_gen(log_order: u32) -> Field64 {
+        // Field64: p = 2^64 - 2^32 + 1, two-adic valuation = 32
+        // TWO_ADIC_ROOT_OF_UNITY has order 2^32
+        let max_log = Field64::TWO_ADICITY;
+        assert!(log_order <= max_log);
+        let mut gen = Field64::TWO_ADIC_ROOT_OF_UNITY;
+        // Raise to 2^(max_log - log_order) to get order 2^log_order
+        for _ in 0..(max_log - log_order) {
+            gen.square_in_place();
         }
-        let g_hats = (0..num_witness_variables)
-            .map(|j| {
-                (0..half)
-                    .map(|i| F::from((i as u64) + (j as u64) + 5))
-                    .collect::<Vec<_>>()
-            })
-            .collect::<Vec<_>>();
-        let polys = BlindingPolynomials::<F> { m_poly, g_hats };
-        let gamma = F::from(9u64);
-        let rho = F::from(4u64);
-        let eval = polys.evaluate_at(gamma, rho);
+        gen
+    }
 
-        let beq = eq_weights_at_gamma(gamma, rho, num_blinding_variables);
-        let covector = Covector::new(beq);
-        let embedding = Identity::<F>::new();
-        let layout_vectors = polys.layout_vectors();
-
-        assert_eq!(
-            eval.m_eval,
-            covector.evaluate(&embedding, &layout_vectors[0])
-        );
-        for (g_idx, g_eval) in eval.g_hat_evals.iter().enumerate() {
+    #[test]
+    fn dlog_identity_is_zero() {
+        for log_order in 1..=8 {
+            let gen = subgroup_gen(log_order);
             assert_eq!(
-                *g_eval,
-                covector.evaluate(&embedding, &layout_vectors[1 + g_idx])
+                discrete_log_pow2(Field64::ONE, gen, log_order),
+                0,
+                "dlog(1, gen, {log_order}) should be 0"
             );
         }
     }
 
     #[test]
-    fn test_construct_batched_eq_weights_matches_per_gamma_accumulation() {
-        let gammas = [F::from(3u64), F::from(6u64), F::from(10u64)];
-        let rho = F::from(2u64);
-        let tau2 = F::from(5u64);
-        let num_blinding_variables = 3usize;
-        let bevals = gammas
-            .into_iter()
-            .map(|gamma| BlindingEvaluations {
-                gamma,
-                m_eval: F::ZERO,
-                g_hat_evals: vec![],
-            })
-            .collect::<Vec<_>>();
-
-        let batched = construct_batched_eq_weights(&bevals, rho, tau2, num_blinding_variables);
-        let mut expected = vec![F::ZERO; 1 << (num_blinding_variables + 1)];
-        let mut power = F::ONE;
-        for gamma in gammas {
-            let per_gamma = eq_weights_at_gamma(gamma, rho, num_blinding_variables);
-            for (acc, value) in expected.iter_mut().zip(per_gamma) {
-                *acc += power * value;
-            }
-            power *= tau2;
+    fn dlog_generator_is_one() {
+        for log_order in 1..=8 {
+            let gen = subgroup_gen(log_order);
+            assert_eq!(
+                discrete_log_pow2(gen, gen, log_order),
+                1,
+                "dlog(gen, gen, {log_order}) should be 1"
+            );
         }
-        assert_eq!(batched.vector, expected);
     }
 
     #[test]
-    fn test_combined_claim_matches_opening_side_recomposition() {
-        let num_blinding_variables = 3usize;
-        let num_witness_variables = 2usize;
-        let half = 1usize << num_blinding_variables;
-        let msk = (0..half)
-            .map(|i| F::from((i as u64) + 7))
-            .collect::<Vec<_>>();
-        let g0_hat = (0..half)
-            .map(|i| F::from((5 * i as u64) + 2))
-            .collect::<Vec<_>>();
-        let mut m_poly = vec![F::from(0u64); 2 * half];
-        for i in 0..half {
-            m_poly[2 * i] = g0_hat[i];
-            m_poly[2 * i + 1] = msk[i];
+    fn dlog_known_powers() {
+        let log_order = 4; // group of order 16
+        let gen = subgroup_gen(log_order);
+
+        for i in 0..16usize {
+            let target = gen.pow([i as u64]);
+            assert_eq!(
+                discrete_log_pow2(target, gen, log_order),
+                i,
+                "dlog(gen^{i}, gen, {log_order}) should be {i}"
+            );
         }
-        let g_hats = (0..num_witness_variables)
-            .map(|j| {
-                (0..half)
-                    .map(|i| F::from((i as u64) + (2 * j as u64) + 11))
-                    .collect::<Vec<_>>()
-            })
-            .collect::<Vec<_>>();
-        let polys = BlindingPolynomials::<F> { m_poly, g_hats };
+    }
 
-        let gammas = [F::from(3u64), F::from(9u64), F::from(12u64)];
-        let rho = F::from(4u64);
-        let tau1 = F::from(6u64);
-        let tau2 = F::from(5u64);
-        let blinding_evals = gammas
-            .into_iter()
-            .map(|gamma| polys.evaluate_at(gamma, rho))
-            .collect::<Vec<_>>();
+    #[test]
+    fn dlog_order_1() {
+        // Group of order 2^1 = 2: elements are {1, -1}
+        let gen = subgroup_gen(1);
+        assert_eq!(discrete_log_pow2(Field64::ONE, gen, 1), 0);
+        assert_eq!(discrete_log_pow2(gen, gen, 1), 1);
+    }
 
-        let direct_claim =
-            blinding_evals
-                .iter()
-                .enumerate()
-                .fold(F::ZERO, |acc, (gamma_idx, eval)| {
-                    let tau2_power = tau2.pow([gamma_idx as u64]);
-                    let inner = eval
-                        .g_hat_evals
-                        .iter()
-                        .enumerate()
-                        .fold(F::ZERO, |inner_acc, (j, g_eval)| {
-                            inner_acc + tau1.pow([(j + 1) as u64]) * *g_eval
-                        });
-                    acc + tau2_power * (eval.m_eval + F::from(2u64) * inner)
-                });
-        let (m_claim, g_hat_claims) = compute_per_polynomial_claims(&blinding_evals, tau2);
-        let recomposed_claim = combine_claim_from_components(m_claim, &g_hat_claims, tau1);
-        assert_eq!(direct_claim, recomposed_claim);
+    #[test]
+    fn dlog_larger_group() {
+        let log_order = 10; // group of order 1024
+        let gen = subgroup_gen(log_order);
+
+        // Check a handful of indices including edge cases
+        for &i in &[0, 1, 2, 511, 512, 1023] {
+            let target = gen.pow([i as u64]);
+            assert_eq!(
+                discrete_log_pow2(target, gen, log_order),
+                i,
+                "failed for i={i}"
+            );
+        }
+    }
+
+    // ---------------------------------------------------------------
+    // Property-based tests
+    // ---------------------------------------------------------------
+
+    proptest! {
+        /// Roundtrip: gen^{dlog(gen^i, gen, n)} == gen^i for all i in [0, 2^n).
+        #[test]
+        fn dlog_roundtrip(log_order in 1u32..=12, idx in 0u32..4096) {
+            let order = 1u32 << log_order;
+            let i = (idx % order) as usize;
+            let gen = subgroup_gen(log_order);
+            let target = gen.pow([i as u64]);
+            let result = discrete_log_pow2(target, gen, log_order);
+            prop_assert_eq!(result, i, "dlog roundtrip failed for log_order={}, i={}", log_order, i);
+        }
+
+        /// phi_i_bits output is always < 2^ell (within the valid range).
+        #[test]
+        fn phi_i_bits_in_range(
+            mu in 4usize..=16,
+            ell in 1usize..=4,
+            hypercube_idx_raw in 0usize..65536,
+        ) {
+            prop_assume!(ell <= mu);
+            let rem = mu % ell;
+            let num_phis = 1 + (mu - rem) / ell; // Φ₀ + ⌊(mu-rem)/ell⌋
+            let max_idx = 1usize << mu;
+            let hypercube_idx = hypercube_idx_raw % max_idx;
+
+            for phi_index in 0..num_phis {
+                // Ensure the window fits: start + ell <= mu
+                let start = if phi_index == 0 { 0 } else { (phi_index - 1) * ell + rem };
+                if start + ell > mu {
+                    break;
+                }
+                let result = phi_i_bits(hypercube_idx, phi_index, mu, ell, rem);
+                prop_assert!(
+                    result < (1 << ell),
+                    "phi_i_bits({}, {}, {}, {}, {}) = {} >= 2^{}",
+                    hypercube_idx, phi_index, mu, ell, rem, result, ell
+                );
+            }
+        }
+
+        /// For rem=0 and ell dividing mu evenly, the Φ windows partition the mu bits.
+        /// Concatenating all windows should reconstruct the original index.
+        #[test]
+        fn phi_i_bits_partition_no_remainder(
+            mu_factor in 1usize..=4,
+            ell in 2usize..=4,
+            hypercube_idx_raw in 0usize..65536,
+        ) {
+            let mu = mu_factor * ell;
+            prop_assume!(mu <= 16);
+            let rem = 0;
+            let num_phis = mu / ell; // exactly mu/ell windows when rem=0
+            let max_idx = 1usize << mu;
+            let b = hypercube_idx_raw % max_idx;
+
+            // Φ₀ extracts top ell bits, Φ₁..Φ_{num_phis-1} extract subsequent windows.
+            // But note: when rem=0, Φ₀ and Φ₁ have the SAME start (both 0).
+            // So Φ₀ through Φ_{num_phis} with phi_index 0, 2, 3, ..., num_phis
+            // actually: with rem=0, Φ₁ start = (0)*ell + 0 = 0 (same as Φ₀)
+            // So the distinct windows are Φ₀, Φ₂, Φ₃, ..., Φ_{num_phis}
+            // which is Φ₀ (start=0), Φ₂ (start=ell), Φ₃ (start=2*ell), ...
+
+            // Reconstruct from non-overlapping windows
+            let mut reconstructed = 0usize;
+            // Φ₀: start=0
+            let phi0 = phi_i_bits(b, 0, mu, ell, rem);
+            reconstructed |= phi0 << (mu - ell);
+
+            // Φ_{i+1}: start = i*ell for i >= 1
+            for i in 1..num_phis {
+                let phi_idx = i + 1; // Φ₂, Φ₃, ...
+                let start = (phi_idx - 1) * ell; // + rem = 0
+                if start + ell > mu {
+                    break;
+                }
+                let bits = phi_i_bits(b, phi_idx, mu, ell, rem);
+                let shift = mu - start - ell;
+                reconstructed |= bits << shift;
+            }
+
+            prop_assert_eq!(
+                reconstructed, b,
+                "partition reconstruction failed for b={:#b}, mu={}, ell={}", b, mu, ell
+            );
+        }
     }
 }
